requirements.txt
Flask==2.3.3
langchain==0.0.287
openai==0.28.1
python-dotenv==1.0.0
